Refactor: move apk information to a separate file into Object/
