+++ quark/main.py
@@ -2,12 +2,11 @@ import argparse
 import operator
 import os
 
-from androguard.misc import AnalyzeAPK
 from prettytable import PrettyTable
 from tqdm import tqdm
 
 from quark.Evaluator.pyeval import PyEval
-from quark.Objects.BytecodeObject import BytecodeObject
+from quark.Objects.Apkinfo import Apkinfo
 from quark.Objects.RuleObject import RuleObject
 from quark.logo import logo
 from quark.utils.colors import (
@@ -19,20 +18,16 @@ from quark.utils.colors import (
     COLOR_OUTPUT_GREEN,
 )
 from quark.utils.out import print_success, print_info, print_warning
-from quark.utils.tools import remove_dup_list
 from quark.utils.weight import Weight
 
 MAX_SEARCH_LAYER = 3
 CHECK_LIST = "".join(["\t[" + u"\u2713" + "]"])
 
+
 class XRule:
     def __init__(self, apk):
 
-        self.a, self.d, self.dx = AnalyzeAPK(apk)
-
-        # Create Class, Method, String and Field
-        # crossreferences for all classes in the Analysis.
-        # self.dx.create_xref()
+        self.apkinfo = Apkinfo(apk)
 
         self.pre_method0 = []
         self.pre_method1 = []
@@ -50,106 +45,6 @@ class XRule:
         # Sum of the each rule
         self.score_sum = 0
 
-    @property
-    def permissions(self):
-        """
-        :returns: A list of permissions
-        :rtype: list
-        """
-        return self.a.get_permissions()
-
-    def find_method(self, class_name=".*", method_name=".*"):
-        """
-        Find method from given class_name and method_name,
-        default is find all.
-
-        :returns: an generator of MethodClassAnalysis
-        :rtype: generator
-        """
-
-        result = self.dx.find_methods(class_name, method_name)
-
-        if len(list(result)) > 0:
-            return self.dx.find_methods(class_name, method_name)
-
-        else:
-            # Method Not Found
-            return None
-
-    def upperFunc(self, class_name, method_name):
-        """
-        Return the upper level method from given class name and
-        method name.
-        :param class_name:
-        :param method_name:
-        :return: list
-        """
-
-        result = []
-        method_set = self.find_method(class_name, method_name)
-
-        if method_set is not None:
-            for md in method_set:
-                for _, call, _ in md.get_xref_from():
-                    # Get class name and method name:
-                    # call.class_name, call.name
-                    result.append((call.class_name, call.name))
-
-            return remove_dup_list(result)
-        else:
-            return None
-
-    def get_method_bytecode(self, class_name, method_name):
-        """
-        Return the corresponding bytecode according to the
-        given class name and method name.
-        :param class_name:
-        :param method_name:
-        :return: generator
-        """
-
-        result = self.dx.find_methods(class_name, method_name)
-
-        if len(list(result)) > 0:
-            for m in self.dx.find_methods(class_name, method_name):
-                for idx, ins in m.get_method().get_instructions_idx():
-                    bytecode_obj = None
-                    reg_list = []
-
-                    # count the number of the registers.
-                    length_operands = len(ins.get_operands())
-                    if length_operands == 0:
-                        # No register, no parm
-                        bytecode_obj = BytecodeObject(ins.get_name(), None, None)
-                    elif length_operands == 1:
-                        # Only one register
-
-                        reg_list.append(
-                            "v" + str(ins.get_operands()[length_operands - 1][1])
-                        )
-                        bytecode_obj = BytecodeObject(ins.get_name(), reg_list, None,)
-                    elif length_operands >= 2:
-                        # the last one is parm, the other are registers.
-
-                        parameter = ins.get_operands()[length_operands - 1]
-                        for i in range(0, length_operands - 1):
-                            reg_list.append("v" + str(ins.get_operands()[i][1]))
-                        if len(parameter) == 3:
-                            # method or value
-                            parameter = parameter[2]
-                        else:
-                            # Operand.OFFSET
-                            parameter = parameter[1]
-
-                        bytecode_obj = BytecodeObject(
-                            ins.get_name(), reg_list, parameter,
-                        )
-
-                    yield bytecode_obj
-        else:
-            return None
-            # TODO add logging
-
     def find_f_previous_method(self, base, top):
         """
         Find the previous method based on base method
@@ -160,8 +55,7 @@ class XRule:
         :param top:
         :return: None
         """
-
-        method_set = self.upperFunc(base[0], base[1])
+        method_set = self.apkinfo.upperFunc(base[0], base[1])
 
         if method_set is not None:
 
@@ -185,7 +79,7 @@ class XRule:
         :return: None
         """
 
-        method_set = self.upperFunc(base[0], base[1])
+        method_set = self.apkinfo.upperFunc(base[0], base[1])
 
         if method_set is not None:
             if top in method_set:
@@ -226,17 +120,16 @@ class XRule:
                 next_list1 = []
                 next_list2 = []
                 for item in list1:
-                    if self.upperFunc(item[0], item[1]) is not None:
-                        next_list1.extend(self.upperFunc(item[0], item[1]))
+                    if self.apkinfo.upperFunc(item[0], item[1]) is not None:
+                        next_list1.extend(self.apkinfo.upperFunc(item[0], item[1]))
                 for item in list2:
-                    if self.upperFunc(item[0], item[1]) is not None:
-                        next_list2.extend(self.upperFunc(item[0], item[1]))
+                    if self.apkinfo.upperFunc(item[0], item[1]) is not None:
+                        next_list2.extend(self.apkinfo.upperFunc(item[0], item[1]))
                 # Append first layer into next layer
                 for pre_list in list1:
                     next_list1.append(pre_list)
                 for pre_list in list2:
                     next_list2.append(pre_list)
-
 
                 # To find the same method, push the previous two
                 # methods into the stack here. Once it found there
@@ -258,7 +151,7 @@ class XRule:
         :return: boolean
         """
 
-        method_set = self.find_method(same_method[0], same_method[1])
+        method_set = self.apkinfo.find_method(same_method[0], same_method[1])
         seq_table = []
 
         if method_set is not None:
@@ -312,8 +205,8 @@ class XRule:
         # Check if there is an operation of the same register
         state = False
 
-        for bytecode_obj in self.get_method_bytecode(
-            common_method[0], common_method[1]
+        for bytecode_obj in self.apkinfo.get_method_bytecode(
+                common_method[0], common_method[1]
         ):
             # ['new-instance', 'v4', Lcom/google/progress/SMSHelper;]
             instruction = [bytecode_obj.mnemonic]
@@ -347,25 +240,25 @@ class XRule:
         """
 
         # Level 1
-        if set(rule_obj.x1_permission).issubset(set(self.permissions)):
+        if set(rule_obj.x1_permission).issubset(set(self.apkinfo.permissions)):
             rule_obj.check_item[0] = True
 
         # Level 2
         test_md0 = rule_obj.x2n3n4_comb[0]["method"]
         test_cls0 = rule_obj.x2n3n4_comb[0]["class"]
-        if self.find_method(test_cls0, test_md0) is not None:
+        if self.apkinfo.find_method(test_cls0, test_md0) is not None:
             rule_obj.check_item[1] = True
             # Level 3
             test_md1 = rule_obj.x2n3n4_comb[1]["method"]
             test_cls1 = rule_obj.x2n3n4_comb[1]["class"]
-            if self.find_method(test_cls1, test_md1) is not None:
+            if self.apkinfo.find_method(test_cls1, test_md1) is not None:
                 rule_obj.check_item[2] = True
 
                 # Level 4
                 # [('class_a','method_a'),('class_b','method_b')]
                 # Looking for the first layer of the upperfunction
-                upperfunc0 = self.upperFunc(test_cls0, test_md0)
-                upperfunc1 = self.upperFunc(test_cls1, test_md1)
+                upperfunc0 = self.apkinfo.upperFunc(test_cls0, test_md0)
+                upperfunc1 = self.apkinfo.upperFunc(test_cls1, test_md1)
 
                 same = self.find_intersection(upperfunc0, upperfunc1)
                 if same is not None:
@@ -388,7 +281,7 @@ class XRule:
 
                             # Level 5
                             if self.check_parameter(
-                                common_method, str(pre_0[1]), str(pre_1[1])
+                                    common_method, str(pre_0[1]), str(pre_1[1])
                             ):
                                 rule_obj.check_item[4] = True
                                 self.same_operation.append(common_method)
